Refactored function.rs to module function
