+++ src/extension/postgres/function.rs
@@ -1,5 +1,12 @@
-use crate::{ColumnType, DynIden, Expr, IntoIden};
-use crate::{QueryBuilder, QuotedBuilder, SqlWriter};
+use crate::{
+    QueryBuilder, QuotedBuilder, SqlWriter,
+};
+
+pub use create::*;
+pub use drop::*;
+
+pub(crate) mod create;
+pub(crate) mod drop;
 
 /// Creates a new "CREATE or DROP FUNCTION" statement for PostgreSQL.
 #[derive(Debug, Default, Clone, PartialEq, Eq)]
@@ -51,196 +58,6 @@ impl PgFunctionStmt {
     /// ```
     pub fn drop() -> FunctionDropStatement {
         FunctionDropStatement::new()
-    }
-}
-
-/// Represents PostgreSQL function argument modes
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FunctionArgMode {
-    In,
-    Out,
-    Inout,
-    Variadic,
-}
-
-/// Represents a PostgreSQL function argument
-#[derive(Debug, Clone, PartialEq)]
-pub struct FunctionArg {
-    pub(crate) mode: Option<FunctionArgMode>,
-    pub(crate) name: Option<DynIden>,
-    pub(crate) arg_type: ColumnType,
-    pub(crate) default: Option<Expr>,
-}
-
-impl FunctionArg {
-    /// Create a new function argument with a type
-    pub fn new(arg_type: ColumnType) -> Self {
-        Self {
-            mode: None,
-            name: None,
-            arg_type,
-            default: None,
-        }
-    }
-
-    /// Set the argument mode (IN, OUT, INOUT, VARIADIC)
-    pub fn mode(mut self, mode: FunctionArgMode) -> Self {
-        self.mode = Some(mode);
-        self
-    }
-
-    /// Set the name of the argument
-    pub fn name<T: IntoIden>(mut self, name: T) -> Self {
-        self.name = Some(name.into_iden());
-        self
-    }
-
-    /// Set the default expression for the argument
-    pub fn default<T: Into<Expr>>(mut self, expr: T) -> Self {
-        self.default = Some(expr.into());
-        self
-    }
-}
-
-/// Represents PostgreSQL function behavior/volatility options
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum FunctionBehavior {
-    Immutable,
-    Stable,
-    Volatile,
-    CalledOnNullInput,
-    ReturnsNullOnNullInput,
-    Strict,
-    SecurityInvoker,
-    SecurityDefiner,
-    ParallelUnsafe,
-    ParallelRestricted,
-    ParallelSafe,
-}
-
-/// Represents the return type of a PostgreSQL function
-#[derive(Debug, Clone, PartialEq)]
-pub enum FunctionReturns {
-    Type(ColumnType),
-    Table(Vec<(DynIden, ColumnType)>),
-}
-
-/// Creates a new "CREATE FUNCTION" statement for PostgreSQL
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct FunctionCreateStatement {
-    pub(crate) name: Option<DynIden>,
-    pub(crate) or_replace: bool,
-    pub(crate) args: Vec<FunctionArg>,
-    pub(crate) returns: Option<FunctionReturns>,
-    pub(crate) language: Option<DynIden>,
-    pub(crate) behavior: Vec<FunctionBehavior>,
-    pub(crate) as_definition: Option<String>,
-    pub(crate) sql_body: Option<String>,
-}
-
-impl FunctionCreateStatement {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Set the function name
-    pub fn name<T: IntoIden>(&mut self, name: T) -> &mut Self {
-        self.name = Some(name.into_iden());
-        self
-    }
-
-    /// Use "OR REPLACE" in the CREATE FUNCTION statement
-    pub fn or_replace(&mut self) -> &mut Self {
-        self.or_replace = true;
-        self
-    }
-
-    /// Add an argument to the function
-    pub fn arg(&mut self, arg: FunctionArg) -> &mut Self {
-        self.args.push(arg);
-        self
-    }
-
-    /// Add multiple arguments to the function
-    pub fn args<I: IntoIterator<Item = FunctionArg>>(&mut self, args: I) -> &mut Self {
-        self.args.extend(args);
-        self
-    }
-
-    /// Set the return type
-    pub fn returns(&mut self, returns: FunctionReturns) -> &mut Self {
-        self.returns = Some(returns);
-        self
-    }
-
-    /// Set the function language (e.g., PL/pgSQL, SQL)
-    pub fn language<T: IntoIden>(&mut self, lang: T) -> &mut Self {
-        self.language = Some(lang.into_iden());
-        self
-    }
-
-    /// Add a behavior modifier (e.g., IMMUTABLE, STRICT, SECURITY DEFINER)
-    pub fn behavior(&mut self, behavior: FunctionBehavior) -> &mut Self {
-        self.behavior.push(behavior);
-        self
-    }
-
-    /// Set function definition string (AS '...')
-    pub fn as_definition<T: Into<String>>(&mut self, definition: T) -> &mut Self {
-        self.as_definition = Some(definition.into());
-        self
-    }
-
-    /// Set SQL function body
-    pub fn sql_body<T: Into<String>>(&mut self, body: T) -> &mut Self {
-        self.sql_body = Some(body.into());
-        self
-    }
-}
-
-/// Creates a new "DROP FUNCTION" statement for PostgreSQL
-#[derive(Debug, Default, Clone, PartialEq)]
-pub struct FunctionDropStatement {
-    pub(crate) name: Option<DynIden>,
-    pub(crate) if_exists: bool,
-    pub(crate) arg_types: Option<Vec<ColumnType>>,
-    pub(crate) cascade: bool,
-    pub(crate) restrict: bool,
-}
-
-impl FunctionDropStatement {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Set the function name to drop
-    pub fn name<T: IntoIden>(&mut self, name: T) -> &mut Self {
-        self.name = Some(name.into_iden());
-        self
-    }
-
-    /// Use "IF EXISTS" on the DROP FUNCTION statement
-    pub fn if_exists(&mut self) -> &mut Self {
-        self.if_exists = true;
-        self
-    }
-
-    /// Specify the argument types to uniquely identify the function overload to drop
-    pub fn arg_types<I: IntoIterator<Item = ColumnType>>(&mut self, types: I) -> &mut Self {
-        self.arg_types = Some(types.into_iter().collect());
-        self
-    }
-
-    /// Use "CASCADE" on the DROP FUNCTION statement
-    pub fn cascade(&mut self) -> &mut Self {
-        self.cascade = true;
-        self
-    }
-
-    /// Use "RESTRICT" on the DROP FUNCTION statement
-    pub fn restrict(&mut self) -> &mut Self {
-        self.restrict = true;
-        self
     }
 }
 
@@ -302,7 +119,7 @@ impl_function_statement_builder!(FunctionDropStatement, prepare_function_drop_st
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{Alias, ColumnType};
+    use crate::{Alias, ColumnType, IntoIden};
 
     // ── FunctionArg ──────────────────────────────────────────────────────────
